restructure folders in server side
