+++ server/src/interfaces/basket.interface.ts
