+++ server/src/interfaces/product.interface.ts
