+++ server/src/interfaces/user.interface.ts
