+++ server/src/routes/product.routes.ts
@@ -1,6 +1,6 @@
 import * as express from "express";
 import { ObjectId } from "mongodb";
-import { collections } from "./database";
+import { collections } from "../database";
 
 export const productRouter = express.Router();
 productRouter.use(express.json());
