+++ server/src/server.ts
@@ -2,9 +2,9 @@ import * as dotenv from "dotenv";
 import express from "express";
 import cors from "cors";
 import { connectToDatabase } from "./database";
-import { productRouter } from "./product.routes";
-import { userRouter } from "./user.routes";
-import { basketRouter } from "./basket.routes";
+import { productRouter } from "./routes/product.routes";
+import { userRouter } from "./routes/user.routes";
+import { basketRouter } from "./routes/basket.routes";
 import passport from './auth';
 import session from 'express-session';
 import MongoStore from "connect-mongo";
